+++ Variables/ComponentNumber.cs
@@ -4,6 +4,6 @@ namespace Screenplay.Variables
 {
     public abstract class ComponentNumber<T> : ComponentVariable<T>, INumber where T : IComparable<T>, IConvertible
     {
-        public double GetNumber() => Value.ToDouble(null);
+        public decimal GetNumber() => Value.ToDecimal(null);
     }
 }
