+++ Variables/IBool.cs
@@ -2,6 +2,6 @@ namespace Screenplay.Variables
 {
     public interface IBool : IValue<bool>, INumber
     {
-        double INumber.GetNumber() => Value ? 1.0 : 0.0;
+        decimal INumber.GetNumber() => Value ? 1 : 0;
     }
 }
