+++ Variables/IFloat.cs
@@ -2,6 +2,6 @@ namespace Screenplay.Variables
 {
     public interface IFloat : IValue<float>, INumber
     {
-        double INumber.GetNumber() => Value;
+        decimal INumber.GetNumber() => (decimal)Value;
     }
 }
