+++ Variables/IInt.cs
@@ -2,6 +2,6 @@ namespace Screenplay.Variables
 {
     public interface IInt : IValue<int>, INumber
     {
-        double INumber.GetNumber() => Value;
+        decimal INumber.GetNumber() => Value;
     }
 }
