+++ Variables/INumber.cs
@@ -2,6 +2,6 @@ namespace Screenplay.Variables
 {
     public interface INumber : IValue
     {
-        double GetNumber();
+        decimal GetNumber();
     }
 }
